first

+++ app/src/main/java/com/wd/tech/bean/consult/ConsultDetailsBean.java
@@ -75,7 +75,7 @@ public class ConsultDetailsBean {
         private String title;
         private int whetherCollection;
         private int whetherGreat;
-        private int yuanCost;
+        private double yuanCost;
         private List<InformationListBean> informationList;
         private List<PlateBean> plate;
 
@@ -199,11 +199,11 @@ public class ConsultDetailsBean {
             this.whetherGreat = whetherGreat;
         }
 
-        public int getYuanCost() {
+        public double getYuanCost() {
             return yuanCost;
         }
 
-        public void setYuanCost(int yuanCost) {
+        public void setYuanCost(double yuanCost) {
             this.yuanCost = yuanCost;
         }
